resort main-settings categories, move 'read receipts' and 'blocked contacts' to 'chats&media' (communication on ios)
