+++ res/xml/preferences_app_protection.xml
@@ -1,37 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <PreferenceCategory android:title="@string/pref_communication">
+    <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:key="pref_screen_security"
+        android:summary="@string/pref_screen_security_explain"
+        android:title="@string/pref_screen_security" />
 
-        <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:key="pref_read_receipts"
-            android:summary="@string/pref_read_receipts_explain"
-            android:title="@string/pref_read_receipts" />
-
-        <Preference
-            android:key="preference_category_blocked"
-            android:title="@string/pref_blocked_contacts" />
-
-    </PreferenceCategory>
-
-
-    <PreferenceCategory android:title="@string/pref_app_access">
-
-        <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:key="pref_screen_security"
-            android:summary="@string/pref_screen_security_explain"
-            android:title="@string/pref_screen_security" />
-
-        <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
-            android:defaultValue="false"
-            android:key="pref_incognito_keyboard"
-            android:summary="@string/pref_incognito_keyboard_explain"
-            android:title="@string/pref_incognito_keyboard" />
-
-    </PreferenceCategory>
-
+    <org.thoughtcrime.securesms.components.SwitchPreferenceCompat
+        android:defaultValue="false"
+        android:key="pref_incognito_keyboard"
+        android:summary="@string/pref_incognito_keyboard_explain"
+        android:title="@string/pref_incognito_keyboard" />
 
     <PreferenceCategory android:title="@string/screenlock_title">
